+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Next version (not yet released)
 -------------------------------
 
-
+- Use common package cache dir for all grocker instances
 
 Grocker 2.1.0 (2016-02-11)
 --------------------------